+++ CHANGELOG.md
@@ -31,6 +31,10 @@
   affected resource. (https://github.com/pulumi/pulumi-kubernetes/pull/3128)
 - `PersistentVolumeClaims` with a bind mode of `WaitForFirstConsumer` will no
   longer hang indefinitely. (https://github.com/pulumi/pulumi-kubernetes/pull/3130)
+- `Services` with selectors targeting 0 `Pods` will no longer hang indefinitely.
+  (https://github.com/pulumi/pulumi-kubernetes/issues/605)
+- `Services` without selectors will no longer hang indefinitely.
+  (https://github.com/pulumi/pulumi-kubernetes/issues/799)
 
 ## 4.15.0 (July 9, 2024)